requirements.txt
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.30.0,<4.40.0
huggingface-hub>=0.20.0,<0.22.0
pyannote.core>=5.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.1
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0
requests>=2.31.0
# NVIDIA NeMo for diar_sortformer model (optional but recommended)
# Install with: pip install git+https://github.com/NVIDIA/NeMo.git@main#egg=nemo_toolkit[asr]
# Note: Requires Cython and packaging: pip install Cython packaging

